remove code and add tests
