+++ components/function-controller/cmd/webhook/main.go
@@ -117,14 +117,8 @@ func main() {
 	whs.CertName = resources.CertFile
 	whs.KeyName = resources.KeyFile
 
-	defaultCfg, err := webhookCfg.ToDefaultingConfig()
-	if err != nil {
-		setupLog.Error(err, "while creating of defaulting configuration")
-		os.Exit(1)
-	}
 	whs.Register(resources.FunctionDefaultingWebhookPath, &ctrlwebhook.Admission{
 		Handler: webhook.NewDefaultingWebhook(
-			&defaultCfg,
 			mgr.GetClient(),
 			logWithCtx.Named("defaulting-webhook")),
 	})
