+++ components/function-controller/pkg/apis/serverless/v1alpha2/function_defaults.go
@@ -1,11 +1,5 @@
 package v1alpha2
 
-import (
-	"k8s.io/apimachinery/pkg/api/resource"
-
-	corev1 "k8s.io/api/core/v1"
-)
-
 const DefaultingConfigKey = "defaulting-config"
 
 type ResourcesPreset struct {
@@ -38,120 +32,4 @@ type DefaultingConfig struct {
 	Function FunctionDefaulting
 	BuildJob BuildJobDefaulting
 	Runtime  Runtime
-}
-
-func (fn *Function) Default(config *DefaultingConfig) {
-}
-
-func (spec *FunctionSpec) defaultFunctionResources(config *DefaultingConfig, fn *Function) {
-	var resources *corev1.ResourceRequirements
-	var profile string
-	if spec.ResourceConfiguration != nil && spec.ResourceConfiguration.Function != nil {
-		functionResourceCfg := *spec.ResourceConfiguration.Function
-		if functionResourceCfg.Resources != nil {
-			resources = functionResourceCfg.Resources
-		}
-		profile = functionResourceCfg.Profile
-	}
-	defaultingConfig := config.Function.Resources
-	calculatedResources := calculateResources(fn, resources, profile, FunctionResourcesPresetLabel, defaultingConfig.Presets, defaultingConfig.DefaultPreset, defaultingConfig.RuntimePresets)
-	setFunctionResources(spec, calculatedResources)
-}
-
-func setFunctionResources(spec *FunctionSpec, resources *corev1.ResourceRequirements) {
-
-	if spec.ResourceConfiguration == nil {
-		spec.ResourceConfiguration = &ResourceConfiguration{}
-	}
-
-	if spec.ResourceConfiguration.Function == nil {
-		spec.ResourceConfiguration.Function = &ResourceRequirements{}
-	}
-
-	spec.ResourceConfiguration.Function.Resources = resources
-}
-
-func (spec *FunctionSpec) defaultBuildResources(config *DefaultingConfig, fn *Function) {
-	// if build resources are not set by the user we don't default them.
-	// However, if only a part is set or the preset label is set, we should correctly set missing defaults.
-	if shouldSkipBuildResourcesDefault(fn) {
-		return
-	}
-
-	var buildResourceCfg ResourceRequirements
-	if spec.ResourceConfiguration != nil && spec.ResourceConfiguration.Build != nil {
-		buildResourceCfg = *spec.ResourceConfiguration.Build
-	}
-
-	defaultingConfig := config.BuildJob.Resources
-	calculatedResources := calculateResources(fn, buildResourceCfg.Resources, buildResourceCfg.Profile, BuildResourcesPresetLabel, defaultingConfig.Presets, defaultingConfig.DefaultPreset, nil)
-
-	setBuildResources(spec, calculatedResources)
-}
-
-func setBuildResources(spec *FunctionSpec, resources *corev1.ResourceRequirements) {
-
-	if spec.ResourceConfiguration == nil {
-		spec.ResourceConfiguration = &ResourceConfiguration{}
-	}
-
-	if spec.ResourceConfiguration.Build == nil {
-		spec.ResourceConfiguration.Build = &ResourceRequirements{}
-	}
-
-	spec.ResourceConfiguration.Build.Resources = resources
-}
-
-func shouldSkipBuildResourcesDefault(fn *Function) bool {
-	resourceCfg := fn.Spec.ResourceConfiguration.Build
-	_, hasPresetLabel := fn.Labels[BuildResourcesPresetLabel]
-	if hasPresetLabel {
-		return false
-	}
-
-	if resourceCfg != nil {
-		if resourceCfg.Profile != "" {
-			return false
-		}
-		if resourceCfg.Resources != nil {
-			return resourceCfg.Resources.Limits == nil && resourceCfg.Resources.Requests == nil
-		}
-	}
-	return true
-}
-
-func calculateResources(fn *Function, resourceRequirements *corev1.ResourceRequirements, profile string, presetLabel string, presets map[string]ResourcesPreset, defaultPreset string, runtimePreset map[string]string) *corev1.ResourceRequirements {
-	// profile has the highest priority
-	preset := profile
-	// we can use profile from label (deprecated) instead of new profile
-	if preset == "" {
-		preset = fn.GetLabels()[presetLabel]
-	}
-	if preset != "" {
-		return presetsToRequirements(presets[preset])
-	}
-	// when no profile we use user defined resources
-	if resourceRequirements != nil {
-		return resourceRequirements
-	}
-	// we use default preset only when no profile and no resources
-	rtmPreset, ok := runtimePreset[string(fn.Spec.Runtime)]
-	if ok {
-		return presetsToRequirements(presets[rtmPreset])
-	}
-	return presetsToRequirements(presets[defaultPreset])
-}
-
-func presetsToRequirements(preset ResourcesPreset) *corev1.ResourceRequirements {
-	result := corev1.ResourceRequirements{
-		Limits: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse(preset.LimitCPU),
-			corev1.ResourceMemory: resource.MustParse(preset.LimitMemory),
-		},
-		Requests: corev1.ResourceList{
-			corev1.ResourceCPU:    resource.MustParse(preset.RequestCPU),
-			corev1.ResourceMemory: resource.MustParse(preset.RequestMemory),
-		},
-	}
-	return &result
 }
